persist library panels in responsive grid
